requirements.txt
# Test Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-ordering==0.6
pytest-rerunfailures==12.0

# Selenium and Browser Automation
selenium==4.15.2
playwright==1.39.0

# Environment and Configuration
python-dotenv==1.0.0

# Logging and Reporting
allure-pytest==2.13.2
allure-python-commons==2.13.2

# Utilities
requests==2.31.0
PyYAML==6.0.1

# Code Quality
black==23.11.0
flake8==6.1.0
pytest-cov==4.1.0